% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_data.R
\name{format_data}
\alias{format_data}
\title{Format GWAS summary data.}
\usage{
format_data(
  dat,
  snps.merge = w_hm3.snplist,
  snps.remove = MHC.SNPs,
  snp_col = "SNP",
  b_col = "b",
  or_col = "or",
  se_col = "se",
  freq_col = "freq",
  A1_col = "effect_allele",
  A2_col = "other_allele",
  p_col = "pval",
  ncase_col = "ncase",
  ncontrol_col = "ncontrol",
  n_col = "N",
  n = NULL,
  z_col = "z",
  info_col = "info",
  log_pval = FALSE,
  n_qc = F,
  min_freq = 0.01
)
}
\arguments{
\item{dat}{Data frame. Must have header with at least SNP A1 A2 signed statistics pvalue and sample size.}

\item{snps.merge}{Data frame with SNPs to extract. must have headers: SNP A1 and A2. For example, the hapmap3 SNPlist.}

\item{snps.remove}{a set of SNPs needed to be removed. For example, the SNPs in MHC region.}

\item{snp_col}{column with SNP rs IDs. The default is \code{"SNP"}.}

\item{b_col}{Name of column with effect sizes. The default is \code{"b"}.}

\item{se_col}{Name of column with standard errors. The default is \code{"se"}.}

\item{freq_col}{Name of column with effect allele frequency. The default is \code{"freq"}.}

\item{A1_col}{Name of column with effect allele. Must contain only the characters "A", "C", "T" or "G". The default is \code{"A1"}.}

\item{A2_col}{Name of column with non effect allele. Must contain only the characters "A", "C", "T" or "G". The default is \code{"A2"}.}

\item{p_col}{Name of column with p-value. The default is \code{"pval"}.}

\item{ncase_col}{Name of column with number of cases. The default is \code{"ncase"}.}

\item{ncontrol_col}{Name of column with number of controls. The default is \code{"ncontrol"}.}

\item{n_col}{Name of column with sample size. The default is \code{"n"}.}

\item{n}{Sample size}

\item{z_col}{Name of column with Zscore. The default is \code{"z"}.}

\item{info_col}{Name of column with inputation Info. The default is \code{"info_col"}.}

\item{log_pval}{The pval is -log10(p_col). The default is \code{FALSE}.}

\item{n_qc}{Whether to remove SNPs according to the sample size of SNPs. The default is \code{FALSE}.}

\item{min_freq}{SNPs with allele frequecy less than min_freq will be removed.The default is \code{0.01}}

\item{or_col:}{Name of column with odds ratio. The default is \code{or}.}
}
\value{
data frame wih headers: SNP: rsid; A1: effect allele; A2: non effect allel;
Z: Z score;  N: sample size;  chi2: chi square statistics;  P: p-value.
}
\description{
Reads in GWAS summary data. Infer Zscores from p-values and signed satatistics.
This function is adapted from the format_data() function in MRCIEU/TwoSampleMR.
}
