% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRAPPSS.R
\name{MRAPPSS}
\alias{MRAPPSS}
\title{A function for implementing MR-APPSS.}
\usage{
MRAPPSS(
  MRdat = NULL,
  exposure = "exposure",
  outcome = "outcome",
  pi0 = NULL,
  sigma.sq = NULL,
  tau.sq = NULL,
  Sigma_err = matrix(c(1, 0, 0, 1), 2, 2),
  Omega = matrix(0, 2, 2),
  Cor.SelectionBias = T,
  tol = 1e-08,
  ELBO = F
)
}
\arguments{
\item{MRdat}{data frame at least contain the following varaibles: b.exp b.out se.exp se.out L2 Threshold. L2:LD score, Threshold: modified IV selection threshold for correction of selection bias}

\item{exposure}{exposure name}

\item{outcome}{outcome name}

\item{pi0}{initial value for pi0, default `NULL` will use the default initialize procedure.}

\item{sigma.sq}{initial value for sigma.sq , default `NULL`will use the default initialize procedure.}

\item{tau.sq}{initial value for tau.sq , default `NULL` will use the default initialize procedure.}

\item{Sigma_err}{the error term correlation matrix. default `diag(2)`.}

\item{Omega}{the background varaince component. default `matrix(0,2,2)`.}

\item{Cor.SelectionBias}{Whether use the selection Threshold for correction of selection bias. If FALSE, the model won't correct for selection bias.}

\item{tol}{tolerence, default '1e-08'}

\item{ELBO}{Whether check the evidence lower bound or not, if `FALSE`, check the maximum likelihood instead. default `FALSE`.}
}
\value{
a list with the following elements:
\describe{
\item{MRdat: }{Input data frame}
\item{exposure: }{exposure of interest}
\item{outcome: }{outcome of interest}
\item{beta: }{causal effect estimate}
\item{beta.se: }{standard error}
\item{pval: }{p-value}
\item{sigma.sq: }{variance of forground exposure effect}
\item{tau.sq: }{variance of forground outcome effect}
\item{pi0: }{The probability of a SNP with forground signal after selection}
\item{post: }{Posterior estimates of latent varaibles}
\item{method: }{"MR-APPSS"}
}
}
\description{
MR-APPSS: a unified approach to Mendelian Randomization accounting for polygenicity, pleiotropy and sample structure using genome-wide summary statistics.
MA-APPSS uses a variantional EM algorithm for estimation of parameters.
MR-APPSS uses likelihood ratio test for inference.
}
\examples{
library(MRAPPSS)
exposure = "BMI"
outcome = "T2D"
Threshold = 5e-05  # IV selection Threshold
data(Sigma_err)
data(Omega)
data(MRdat)
MRres = MRAPSS(MRdat,
               exposure = "BMI",
               outcome = "T2D",
               Sigma_err = Sigma_err,
               Omega =  Omega ,
               Cor.SelectionBias = T)
MRplot(MRres, exposure = "BMI", outcome = "T2D")
}
