% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRAPSS.R
\name{MRAPSS}
\alias{MRAPSS}
\title{A function for implementing MR-APSS.}
\usage{
MRAPSS(
  MRdat = NULL,
  exposure = "exposure",
  outcome = "outcome",
  pi0 = NULL,
  sigma.sq = NULL,
  tau.sq = NULL,
  Sigma_err = matrix(c(1, 0, 0, 1), 2, 2),
  Omega = matrix(0, 2, 2),
  tol = 1e-08,
  Threshold = 1,
  ELBO = F,
  genome_Threshold = 5e-08
)
}
\arguments{
\item{MRdat}{data frame at least contain the following varaibles: b.exp b.out se.exp se.out L2. L2:LD score}

\item{exposure}{exposure name}

\item{outcome}{outcome name}

\item{pi0}{initial value for pi0, default `NULL` will use the default initialize procedure.}

\item{sigma.sq}{initial value for sigma.sq , default `NULL`will use the default initialize procedure.}

\item{tau.sq}{initial value for tau.sq , default `NULL` will use the default initialize procedure.}

\item{Sigma_err}{the error term correlation matrix. default `diag(2)`.}

\item{Omega}{the background varaince component. default `matrix(0,2,2)`.}

\item{tol}{tolerence, default '1e-08'}

\item{Threshold}{The selection Threshold for correction of selection bias. If Threshold=1, the model won't correct for selection bias.}

\item{ELBO}{Whether check the evidence lower bound or not, if `FALSE`, check the maximum likelihood instead. default `FALSE`.}

\item{genome_Threshold}{Genome threshold suggesting strong association , which will be used in the default initialize procedure. default `5e-08`.}
}
\value{
a list with the following elements:
\describe{
\item{MRdat}{Input data frame}
\item{beta}{causal effect estimate}
\item{beta.se}{standard error}
\item{pval}{p-value}
\item{sigma.sq}{variance of forground exposure effect}
\item{tau.sq}{variance of forground outcome effect}
\item{pi0}{The probability of a SNP with forground signal after selection}
\item{post}{Posterior estimates of latent varaibles}
\item{method}{"MRAPSS"}
}
}
\description{
MR-APSS: a unified approach to Mendelian Randomization accounting for pleiotropy, sample overlap and selection bias  using genome wide summary statistics.
MA-APSS uses a variantional EM algorithm for estimation of parameters.
The algorithm defaultly uses estimates based on strong associations (p<genome_Threshold) to initialize nuisance parameters.
Users can specify the genome_Threshold by themselves.
Users can also specify the initial values for nuisance parameters.
MR-APSS uses likelihood ratio test for inference.
}
\examples{
library(MRAPSS)
exposure = "BMI"
outcome = "T2D"
Threshold = 5e-05  # IV selection Threshold
data(Sigma_err)
data(Omega)
data(MRdat)
MRres = MRAPSS(MRdat,
               exposure="BMI",
               outcome= "T2D",
               Sigma_err = Sigma_err,
               Omega =  Omega ,
               Threshold =  Threshold)
MRplot(MRres, exposure="BMI", outcome="T2D")
}
