% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRplot.R
\name{MRplot}
\alias{MRplot}
\title{Visualize the MRAPSS results}
\usage{
MRplot(MRres, exposure = "trait 1", outcome = "trait 2")
}
\arguments{
\item{outcome}{: outcome name}

\item{MRres:}{MRAPSS fit results}

\item{exposure:}{exposure name}
}
\value{
Plot of SNP-exposure effect and SNP-outcome effect with the causal effect and 95\% confidence interval.
}
\description{
Visualize the MRAPSS results
}
